+++ src/re2_cache.h
@@ -7,6 +7,6 @@
  * Look up or compile a pattern. Returns cached re2_pattern.
  * On compile error, returns NULL, errbuf filled.
  */
-re2_pattern *re2_cache_lookup(const char *pattern, int pattern_len, char *errbuf, size_t errbuf_size);
+re2_pattern *re2_cache_lookup(const char *pattern, size_t pattern_len, char *errbuf, size_t errbuf_size);
 
 #endif /* RE2_CACHE_H */
